208-ninja_error_handler/pyproject.toml
[tool.poetry]
name = "208-ninja-error-handler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "208_ninja_error_handler"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-ninja = "^0.22.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
